# claroz-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A RESTful API for the Claroz social platform",
        "title": "Claroz API",
        "termsOfService": "https://claroz.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://claroz.com/support",
            "email": "lucas.little@claroz.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://claroz.com/license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "schemes": ["http"],
    "paths": {
...
        "/posts": {
            "get": {
                "description": "Retrieve a list of posts with pagination support",
                "produces": ["application/json"],
                "tags": ["posts"],
                "summary": "Get posts with pagination",
                "parameters": [{
                    "type": "integer",
                    "description": "Page number (default: 1)",
                    "name": "page",
                    "in": "query",
                    "minimum": 1
                }, {
                    "type": "integer",
                    "description": "Page size (default: 10, max: 50)",
                    "name": "pageSize",
                    "in": "query",
                    "minimum": 1,
                    "maximum": 50
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Post"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [{"Bearer": []}],
                "description": "Create a new post with an image and caption",
                "consumes": ["multipart/form-data"],
                "produces": ["application/json"],
                "tags": ["posts"],
                "summary": "Create a new post",
                "parameters": [{
                    "type": "file",
                    "description": "Image file",
                    "name": "image",
                    "in": "formData",
                    "required": true
                }, {
                    "type": "string",
                    "description": "Post caption",
                    "name": "caption",
                    "in": "formData"
                }],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Post"
                        }
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "description": "Retrieve a single post by its ID",
                "produces": ["application/json"],
                "tags": ["posts"],
                "summary": "Get a post by ID",
                "parameters": [{
                    "type": "string",
                    "description": "Post ID",
                    "name": "id",
                    "in": "path",
                    "required": true
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Post"
                        }
                    }
                }
            },
            "delete": {
                "security": [{"Bearer": []}],
                "description": "Delete a post by ID (only by post owner)",
                "produces": ["application/json"],
                "tags": ["posts"],
                "summary": "Delete a post",
                "parameters": [{
                    "type": "string",
                    "description": "Post ID",
                    "name": "id",
                    "in": "path",
                    "required": true
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MessageResponse"
                        }
                    }
                }
            }
        },
        "/posts/{id}/comments": {
            "post": {
                "security": [{"Bearer": []}],
                "description": "Add a new comment to a specific post",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["posts"],
                "summary": "Add a comment to a post",
                "parameters": [{
                    "type": "string",
                    "description": "Post ID",
                    "name": "id",
                    "in": "path",
                    "required": true
                }, {
                    "description": "Comment content",
                    "name": "comment",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/handlers.CommentRequest"
                    }
                }],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Comment"
                        }
                    }
                }
            }
        },
        "/posts/{id}/like": {
            "post": {
                "security": [{"Bearer": []}],
                "description": "Add a like to a specific post",
                "produces": ["application/json"],
                "tags": ["posts"],
                "summary": "Like a post",
                "parameters": [{
                    "type": "string",
                    "description": "Post ID",
                    "name": "id",
                    "in": "path",
                    "required": true
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [{"Bearer": []}],
                "description": "Remove a like from a specific post",
                "produces": ["application/json"],
                "tags": ["posts"],
                "summary": "Unlike a post",
                "parameters": [{
                    "type": "string",
                    "description": "Post ID",
                    "name": "id",
                    "in": "path",
                    "required": true
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MessageResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/follow": {
            "post": {
                "security": [{"Bearer": []}],
                "description": "Follow another user",
                "produces": ["application/json"],
                "tags": ["users"],
                "summary": "Follow a user",
                "parameters": [{
                    "type": "string",
                    "description": "User ID to follow",
                    "name": "id",
                    "in": "path",
                    "required": true
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [{"Bearer": []}],
                "description": "Unfollow a previously followed user",
                "produces": ["application/json"],
                "tags": ["users"],
                "summary": "Unfollow a user",
                "parameters": [{
                    "type": "string",
                    "description": "User ID to unfollow",
                    "name": "id",
                    "in": "path",
                    "required": true
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CommentRequest": {
            "type": "object",
            "required": ["content"],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Great post!"
                }
            }
        },
        "handlers.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
...
        "models.Post": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "caption": {
                    "type": "string",
                    "example": "Beautiful sunset!"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://claroz.com/images/sunset.jpg"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-26T00:35:27Z"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-26T00:35:27Z"
                }
            }
        },
        "models.Comment": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "post_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "content": {
                    "type": "string",
                    "example": "Great post!"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-26T00:35:27Z"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Type \"Bearer\" followed by a space and JWT token."
        }
    }
}